+++ src/helpers/html/index.ts
@@ -2,3 +2,4 @@ export {default as checkedIf} from '~/helpers/html/checkedIf';
 export {default as classIf} from '~/helpers/html/classIf';
 export {default as options} from '~/helpers/html/options';
 export {default as selectedIf} from '~/helpers/html/selectedIf';
+export {default as stripTags} from '~/helpers/html/stripTags';